fix: Add more descriptive names to SQS timers (#76)
The SQS service times sending messages and logs a message
like sqs-send-message-<UUID> took 40ms.

In services where messages are sent to several queues
at once (e.g. giftaid submit) the current log messages
are insufficient when something goes wrong
e.g. if one send times out. Also, the UUID is randomly
generated just for the timer, but in this context looks
like it should be a message ID, which is misleading.

The full qualified queue URL is not included in the timer id
for each timer where the queue URL is available.