fix: Allow HTTP4xx to be logged on Epsagon
Our error handling on Lambda Wrapper checks whether `error.code` is defined and below 500 (meaning a 4xx error). While this makes sense in most uses cases, as we are unlikely to be interested at all in 4xx errors, there are scenarios where we are forced to throw a 4xx and we might be interested in logging in.

This is correlated to [ENG-110].

Error handling will check whether `error.raiseOnEpsagon` is defined and true. This condition will override the error code check, allowing us to put some dynamic logic on whether an error should be logged on Epsagon regardless of code. For example, we might not care about the first time we get an HTTP 417 and just retry, but be alarmed once it gets on the DLQ.

See:
- [ENG-110]
- [ENG-111]