feat: Define logger.warning function (#124)
I have been thinking on how logger.error gets in the way of testing. I believe failures should be as noisy as they can be, and also that they should be tested.

When it comes to test a function failure that has the side effect of calling logger.error what we really want to observe is some kind of error handling (i.e. the function forces the lambda to return a form of HTTP 4xx). In production usage, we might want to trigger logger.error. This is even more true when the lambda is triggered by SQS. The reason is that we want to log the error to Epsagon while not necessarily fail the function (especially if we are processing several messages in one call).

Enter feature tests. The behaviour above is all good for production, but when running integration tests it spams Epsagon with false positives. Staging functions are full of errors that are triggered on purpose by feature tests (checking that things fail nicely) . This is noise that don't help us on keeping track of real errors that might slip through our attention.

My proposal is to add a logger.warning function that checks the
environment (process.env.LOGGER_SOFT_WARNING) and calls logger.error in
production and logger.info otherwise - This is easier to manage (we
don't need to deal with SQS or downstream failures) but introduces a
difference between staging and production.

See:
- [ENG-125]