feat: Call SQS triggered lambdas directly when using serverless-offline
When using `serverless-offline`, LambdaWrapper performs an HTTP Request trying to pass its content to a message queue instance running on localhost. We can intercept the HTTP Request, but it is problematic to code and to think about.

Serverless Offline allows us to trigger lambdas via the AWS SDK. We can map lambdas to queues via an environment variable and trigger functions directly. This has the added benefit of allowing any number of triggers (i.e. one function that queues 10 messages) and to simplify how we test those functions locally.

See:
- [ENG-245]