Rephrase some things to improve clarity
In particular, it can be quite confusing when we say "lambda" whether
we mean the actual AWS Lambda service (real or offline) or a function
that runs within that service (since we often refer to those as simply
"lambdas"). I've tried to use consistent capitalisation and explicitly
say "function" when we're talking about functions.

Also tidied up the Markdown style.