feat: Get SQS account ID from env when offline
Currently, the SQS account ID (which is required to build a queue URL)
is extracted from function ARN in the Lambda context. When running in
`serverless-offline`, the context is faked and the account ID comes out
as `undefined`, resulting in invalid queue URLs.

This issue happens only when using `serverless-offline` and the `aws`
offline SQS mode.

To solve it, `SQSService` will now fall back to getting the account ID
from an environment variable `AWS_ACCOUNT_ID`.