feat: Migrate to TypeScript (#882)
This is a first iteration of Lambda Wrapper v2, which contains a number
of breaking changes.

The key aims are:

- Migrate to a TypeScript-only codebase
- Reconsider how dependency injection should work so that dependency
  types can be inferred
- Fully document the new version
- Keep other changes to the package's API to a minimum

For a complete list of changes, please read the [v2 migration guide](https://github.com/comicrelief/lambda-wrapper/blob/beta/docs/migration/v2.md).

In terms of what we'll do with this moving forward: v2 will first be
published on the beta release channel. We can then test it out in one
or two real-life applications to see how easy or difficult it is to
work with, and continue to release improvements to the beta channel
until we are satisfied that it's ready for wider use.

BREAKING CHANGE: See the [v2 migration guide](https://github.com/comicrelief/lambda-wrapper/blob/beta/docs/migration/v2.md) for full details.