feat: Export `RESPONSE_HEADERS` and fix its type
Makes `RESPONSE_HEADERS` available for import from the package index.
This is useful if you want to change the default headers in your
responses, e.g. to change `Content-Type`.

The type is also relaxed to `Record<string, string>` to allow adding or
removing headers.

BREAKING CHANGE: the `Access-Control-Allow-Credentials` header value is
now the string `'true'` instead of Boolean `true`.