feat: Throw if there are conflicting dependency names (#1155)
Adds validation to check that all dependencies have a unique name.

This follows on from [a discussion] with @zhibek about why we need to
turn off webpack's minification optimisation. TL;DR two dependency
classes can be minified to the same one-letter name, and since
dependencies are keyed by class name, this causes nasty runtime errors.

Lambda Wrapper will now be able to detect this situation and suggest
turning off minification, linking to the [readme note] of how to do
this for webpack.

As part of this, dependency classes will be deduplicated before being
instantiated. If a dependency is specified several times (unlikely but
possible) it will now be instantiated only once. Noting for completeness
in case we discover unexpected side effects down the line.

Jira: [ENG-2728]

[a discussion]: https://github.com/comicrelief/serverless-payments/pull/614#discussion_r1332757352
[readme note]: https://github.com/comicrelief/lambda-wrapper/tree/beta#notes