Change type of `body` to `unknown`
This is the parsed JSON body, so could have any type. Using `unknown`
because this is safer -- you will have to cast or validate the value
before accessing fields from it.

I thought about adding a type parameter to `SQSMessageModel` for a
generic body type, however without any validation this is not type-safe.
I think we should stick to the `RequestModel` approach of just returning
data and letting the user validate and type it as needed.