refactor: Rename SQS message model class
Change the name of the class to match the name exported from the
package index.

This should not be a breaking change because the class is the default
export of its module. Just in case there is some contrived case that
depends on the class's `name` attribute, we'll include this in the
v2.0.0 major release.