fix: Improve type of `configure` config param
This allows IntelliSense to suggest the `dependencies` key (and any
other existing config keys) when you are entering the config parameter.

The `TMoreConfig` type, which is used to track the addition of arbitrary
keys to the Lambda Wrapper config object, was not constrained and so
could be inferred as `any`.

The type of the `config` param therefore defaulted to
`Partial<TConfig> & any`. This is equivalent to `any`, and IntelliSense
cannot make suggestions about keys that may be available.

By requiring `TMoreConfig` to extend `Partial<LambdaWrapperConfig>`, we
still allow any object but tell the type system about expected optional
keys, so IntelliSense can suggest them.

Jira: [ENG-3163]