test: Add type tests covering `di.get` and `SQSService` (#1198)
First attempt at adding tests for some of the types used in Lambda
Wrapper v2.

In #1196 I tried to fix one problem and inadvertently caused another
more serious problem. I caught this during manual QA, but this is
precisely why automated tests are so valuable. Some of the type-related
features of Lambda Wrapper are complex enough that we should have their
behaviour defined in a test suite.

About these tests: they do not "run", as such, and instead the types are
checked statically by the TypeScript compiler. All assertions happen at
the type system level. Test failures will be flagged by a type error on
the failing assertion. I think it still makes sense to organise these
tests using `describe` and `it` to give structure and context to the
assertions.

The docs for [expect-type](https://github.com/mmkal/expect-type#readme)
are useful. I tried various packages (`ts-expect`, `tsd`) and found
`expect-type` to provide the most expressive API and it doesn't need any
particular setup.

So far I've written tests for only `di.get` and `SQSService`, as this
covers most of the interesting stuff and is where I've been seeing
problems – there's enough here to fail following the changes in #1196.
It would also be good to cover things relating to Lambda Wrapper
configuration, however this was not as straightforward as I hoped so I'm
leaving it for now.

Jira: [ENG-3188]