fix: Simplify promises and callbacks (#1203)
Replaces some of the harder-to-understand use of callback-style async
and `new Promise` with simpler async-await. As a bonus, we can remove
the `async` dependency!

This also fixes a bug in `SQSService#checkStatus`, where a failed
`listQueues` call can result in a `TypeError` if the `data` parameter of
the callback is undefined or null. Previously the unit tests ignored
this by passing an empty object (see the deleted `createCallbackMock`
function).

No specific Jira ticket, but removal of `async` completes [ENG-3212]

(Note at time of merge: a substantial amount of the callback-style
async calls to AWS APIs were already replaced with Promise chains
when we migrated to AWS SDK v3 in #1201. I've preserved the original
commit message, but it is no longer completely accurate!)

[ENG-3212]: https://comicrelief.atlassian.net/browse/ENG-3212