ci: Pin Node version in `publish` job
Not clear why this job wasn't already using `actions/setup-node`, but
seems that Node 20 is now the default.

Setting to Node 18 for consistency with other jobs, for now, so we can
unblock that publish. We do need to migrate everything to Node 20+, but
as the last publish failure demonstrated, there's an incompatibility
with one of the ESLint plugins. We need to resolve this in our ESLint
config before we can upgrade here.